Address comments from Bas
