+++ crates/rattler/src/install/mod.rs
@@ -43,10 +43,7 @@ pub use apple_codesign::AppleCodeSignBehavior;
 pub use driver::InstallDriver;
 use fs_err::tokio as tokio_fs;
 use futures::{stream::FuturesUnordered, FutureExt, StreamExt};
-pub use installer::{
-    installation_result_record_transaction, prefix_record_transaction, InstallationResultRecord,
-    Installer, InstallerError, Reporter,
-};
+pub use installer::{result_record::InstallationResultRecord, Installer, InstallerError, Reporter};
 #[cfg(feature = "indicatif")]
 pub use installer::{
     DefaultProgressFormatter, IndicatifReporter, IndicatifReporterBuilder, Placement,
